create-episode

+++ src/main/java/com/example/movie_backend/dto/evaluation/EvaluationDTO.java
@@ -13,9 +13,9 @@ public class EvaluationDTO {
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
     private Long id;
 
-    private Integer star;
+    private Long star;
 
-    private Integer userId;
+    private Long userId;
 
-    private Integer movieId;
+    private Long movieId;
 }
